Add icon for cluster cost
A cluster spec can include a `costs` section detailing the cost of running
that cluster per hour.  The format is:

```
"costs": {
  "average": {
    "pricePerHour": 1
  },
  "estimated": {
    "pricePerHour": 2
  },
  "max": {
    "pricePerHour": 3
  }
}
```

A card footer icon will be shown for a single cost flavour.  The order of
preference for selecting the flavour is average, then estimated, then max.

The card footer icon will be given sensible default values for the text and
tooltip. In the case of the "max" cost, these are "Max £3 / hour" and "The
maximum cost for this cluster is £3 per hour."

The defaults can be customized as shown below:

```
"costs": {
  "average": {
    "pricePerHour": 1,
    "text": "£1 per hour",
    "tooltip": "Only £1/hour! Get it whilst its hot."
  }
}
```