Pass copy to security page et al
We need a webpack loader to load the copy.  During a production build, we
don't have access to a webpack loader when building flight-common.  So we have
the application load the copy and pass it to the page.

This also makes it simple to have different applications use different copy.