Initial implementation of launch tokens
Flight Launch can now be given a launch token instead of an AWS access
key/secret key pair.  If the token is valid, the cluster will be launched
using the Alces AWS credentials.  The Alces AWS credentials are taken from the
environment.

Currently all tokens other than 'taken' are considered to be valid.