Cluster specs can specify a runtime
If a runtime is set, a "runtime" icon will be displayed in the footer of the
cluster spec card.

The runtime is passed on to the server and from there onto fly.  This sets an
expired tag on the cluster stack.  Fly can be used to list the expired
clusters.