Remove potential attack vector
Previously, the server placed too much trust in the data sent from the client.
This level of trust was appropriate when we only supported launching clusters
on the user's AWS account, but now we support launching clusters on our AWS
account it is no longer appropriate.

The mechanism to fix this is to change the data sent by the client from a
cluster spec to a cluster spec reference.  The cluster spec reference is the
name of the cluster spec and the name of the cluster specs file from which it
is taken.

Fixes #52