Add email hash to stack name
The same hash that is used in the FlightProfileBucket, is added to the stack
name.

This allows different email addresses to use the same cluster name, but
prevents the same email address from having two clusters with the same name
running at the same time.

If we assume that there is a one-to-one mapping of users to email addresses,
this seems like a sensible restriction.