Fix bug with promoting staging to production
Running the command `dokku deploy` does not recombine the app's config with
the image.  This was resulting in the staging app's config being used for the
production app.  In practice this meant that the MAIL_BCC_ADDRESS env variable
was not sent and the emails were not being BCC'd.

The `dokku tags:deploy` command runs both the release phase in addition to the
deploy phase.  This combines the production app's config with the image,
before deploying it.