Prevent duplicate launches of a token/cluster spec
This commit prevents a token from resulting in multiple launches in two ways:

 1. A token can no longer be used to queue multiple launches of cluster specs.
 2. If a cluster launch job is processed multiple times, subsequent processing
    will be ignored.

When a cluster spec is queued for launch its token is updated to `QUEUED`.
This prevents the same token from being used to queue another cluster spec
(modulo a small race condition).

When a queued cluster spec is processed, check that the spec is still valid.
This will prevent (modulo a small race condition) 1) multiple cluster specs
queued with the same token from being launched; 2) the same cluster spec
(active job) from being processed twice.

In production we will be using SQS, via Shoryuken, as the active job queue.
SQS provides a guarantee that every job will be delivered, but it doesn't
provide a guarantee that the job will be delivered only once.

There small race conditions present in both prevention checks, so it is still
possible that a single token could result in multiple cluster launches.
However, these race conditions are sufficiently small that I think it is not
worth the effort to try to remove them.