Abandon all cluster launches that fail
We make the token available for use again, and report to the user that the
cluster launch failed.  As we discover which types of failure we may wish to
retry, we may decide to develop code that will retry the cluster launch and
inform the user that we are going to do so.