Add docker stop timeout config to create-docker.sh
We need to allow the worker time to finish processing any clusters that it is
part way through before shuting it down.  We give it 20 minutes.  If it's
still running after that it will be hard shutdown and the "launched" or
"failed" emails will not be sent for those clusters.