Per-tenant cluster specs
When loading Flight Launch the identifier of a tenant can be provided as
the first path parameter, e.g., `https://launch.alces-flight.com/bigVuni`.

The tenant identifier, if any, is included in the request to load the
cluster specs. It is also provided to the server when requesting a
cluster launch and used by the server to load the cluster specs.

This then provides a very simple means of having per-tenant cluster
specs.