Add branding module
The branding module provides a few useful components for adding
branding.  There are <Header /> and <Logo /> which display the branding
header and branding logo respectively.  There is also the `<WithBranding
/>` component which is useful for bespoke branding requirements.
`WithBranding` will call its function child with the branding object if
any.

Currently, branding is added in the nav bar, the tag line and in the
cluster specs section header.