All tokens require a credit allocation
If the token's tenant does not have a credit limit, the token will be created,
if the tenant does, it will be checked to ensure that it has enough credits
remaining.

This logic allows us to have two distinct use-cases for tenants:

 1. A tenant created for a customer. They need to purchase credits from us.
 2. A tenant created for Alces. There is no need for us to ensure that we have
    enough remaining credits.

In both of these cases, the credits allocated to the token, along with the
selected cluster spec will determine how long the cluster will run for.