Prevent loading multiple copies of react
`flight-reactware` includes React as a dev dependency, but not as a
dependency.  `flight-launch` depends on `flight-reactware`.

Under normal `yarn add flight-reactware` usage, the dev dependencies of
`flight-reactware` would not be included and we would have a single copy
of `react` under `flight-launch`es node_modules.

However, when we `yarn link` `flight-reactware` into `flight-launch` we
will end up with a copy of every installed dependency for
`flight-reactware` including its dev dependencies.  In particular,
multiple copies of `react` under `flight-launch`es node_modules.

This commit ensures that only a single copy of `react` is included when
building `flight-launch`.