Don't drop initial state when using loadingStates
Previously, the initial state set by tenants, tokens and clusterSpecs
reducers was being ignored in favour of the initial state set by the
loadingStates reducer.

The issue was that the loadingStates reducer was being called first with
a state of `undefined` and returning its initial state.  The tenant
reducer was then called for the first time with this initial state.  As
that state was not `undefined`, the tenant reducer's initial state was
not used.

This commit fixes the issue by having `composeReducers` reduce the
reducers from the right.  This result is that the tenants reducer is
called first and returns its initial state, then the loadingStates
reducer is called which now has special handling for merging in its
initial state.