Replace addReducers with combineReducers
The use of addReducers was resulting in warnings being written to the
console by its use of combineReducers.  The warnings were highlighting a
fragile use of combineReducers.  Whilst it was OK in our particular use
its fragility meant that that coudl change.

Now we use combineReducers as it was intened to be used.