Fix routing again
Set the default tenant identifier in the redux reducer, so that we can
query it independently of the initial route into the application.
Without this loading the application on, say the security policy page,
would result in broken links in the product nav bar.

Also, in the parts of the app that require a tenant identifier, it is not
optional.