Extract code to ClusterContext and withCluster HOC
Ideally there would be no need for `withCluster`.  The props that it generates
would ideally be included in `ClusterContext` and passed to the component
rendered by `ClusterContext`.  Unfortunately, `ClusterContext` uses
`renderRoutes`, which does not support being passed any props.  So we also need
a `withCluster` HOC to wrap the components that are eventually rendered inside
of `ClusterContext`.