Fix issue launching cluster specs
Previously, we always declared that the tenant the cluster is being
launched for is the "default" tenant.  This would result in the token
not being found, the cluster spec not being found or the cluster spec
key not matching.

Now, we correctly declare the tenant that the cluster belongs to.