UI to configure a single queue when launching
Screen real estate limits this UI to a single queue.  A better UI is
needed, but real estate will be an issue with the current way that the
form is displayed.  Perhaps we need to display the form in another way
which allows for more real estate?  Such as in a modal dialog.