Add credit usages model and resource
The credit usage resource includes calculated fields for the duration of
the credit usage and the accrued charge for the credit usage.  An
accounting period start and end date times can optionally be given, the
resources returned are restricted to only those in the accounting period
and the those resources have their duration and accrued usage restricted
accordingly.