Only create CreditUsages if cluster consumes credits
Clusters launched with a launch token and those consuming a user's
credits have their lifecycles handled differently.  In particular,
clusters using a token do not count towards any consumption of the
user's credits and should not be terminated when the user's credits
expire.

This commit flags the cluster according to the mechanism used to launch
it and prevents creation of the CreditUsages for such clusters.