Introduce "logic" concept to load users
Modules can now export an array of `logic` functions which are called
each time the redux store is changed.  A logic function is called with
`dispatch` and `getState` from the redux store and can perform "business
logic" in response to the redux store changing.

This is a very simple means of preventing a proliferation of React
components which have lifecylce methods defined to perform business
logic.  In time it might be a good idea to replace this with
`redux-logic` or something else.