Fix issue with poor handling of launch errors
All failures to launch the cluster now have the launch token set back to
status AVAILABLE, and an attempt to send a failed email will be made.

If the failed email is sent, we have notified the user and they can take
appropriate action such as trying again with the same launch token.

If the failed email cannot be sent, for example because there are
network issues, an exception will be raised, and the SQS job will be
attempted some 20 minutes later.  As the token is AVAILABLE, this second
attempt should progress as normal.  If any underlying issues with
sending emails have been resolved, the cluster will launch correctly.

There are still potential issues with handling all possible errors.  One
possible example would be if the cluster is started in cloudformation
correctly, but there are then issues with Launch communicating with its
database or sending the completed email.  If this were to happen, the
cluster would be consuming our AWS credits and the user's compute
credits without the user having received the access details for it.
Quite what the correct solution to each of these possible failures would
be I'm not quite sure.

Also, move logging to a better location.