Fix issue with top up form validations
The underlying issue was to do with `react-router-config`'s
`renderRoutes` and its nesting under `CSSTransitionGroup`.  This was
causing the form to be mounted twice and unmounted once when the top up
page was displayed.

As redux-form's have side effects when mounted and unmounted the result
of this was a broken form.  In particular, the validation was not
working correctly.

The solution was to ensure that `renderRoutes` uses the same `location`
when transitioning out as it used when transitioning in.  This ensures
that it will display the same components when transitioning out. The
`renderRoutes` that is transitioning in will display the new components.