Prevent unnecessary page transitions
`location.key` changes every time a react router link is clicked, even
if the route doesn't change.  This results in a transition every time a
react router link is clicked along with the associated mounting and
unmounting of components.

Not only does this flickering not look nice, there can also be
unintended side-effects from the unnecessary mounting and unmounting of
the components.

Using `location.pathname` does not have this issue.

Allowing the key to be overridden by the route allows certain routes to
opt out of transitions.