Refactor cost section of cluster specs
Previously, cluster specs specified a `costPerHour`, which was used when
taking payment up front for the cluster, either via a token, or via an
account's compute credits.

The `costPerHour` has now been replaced with a `chargingModel`
attribute.  The `chargingModel` can specify both an
`upfront.clusterCostPerHour` and an `ongoing.masterNodeCostPerHour`.
The `upfront.clusterCostPerHour` works identically to the previous
`clusterCostPerHour`.

The `ongoing.masterNodeCostPerHour` is used when an account's credits
are being consumed whilst the cluster is running.  It specifies the cost
of the master node, any compute queues are charged in addition to that.