Fix bug with logging in again
When logging out we do not yet clear the redux state.  This results in
the previously logged in user still being in the redux state.  If we log
in as that user, a promise is not returned from the `loadUser` actions,
which breaks the assumption made by our session logic.

This commit removes that assumption from the session logic.