Correct access redirection for credit usages page
When navigating from another Flight platform app,
`launchUsers.selectors.currentUser` will initially return `null`.  If we
branch on that value, we will incorrectly redirect away from the page.
We can't use `launchUsers.selectors.retrieval` to branch as that the
retrieval will never be initiated if there is no signed in user.

`auth.selectors.currentUserSelector` will return a value when the
application first loads iff there is a signed in user.  So we branch on
that.