Initial work on payment data refactoring
The aim is to record more information about how the cluster was "paid"
for, i.e., a token was used, credits were paid upfront, or credits are
being consumed on an ongoing basis.

The new model also needs to make it easy for the client to support (at
least temporarily) marketplace clusters for which we have no payment
information.