Fix race condition
Detecting which clusters have been terminated suffered from a race
condition.  We obtained the list of live clusters and once that was
available we obtained the list of cluters known to the launch server.

It was possible that a cluster would become live and known to the launch
server between the list of live clusters being obtained and the list of
known clusters being obtained.  This would result in the cluster being
incorrectly marked as terminated.

The race condition is fixed by reversing the order.