Fix possible confusion with grace period emails
If a cluster has reached its compute unit limit, we don't want it to be
processed by the `ReduceUsersCreditsJob`.

The `ReduceUsersCreditsJob` sends an email stating that adding compute
units to the account will "save" the cluster.  That will not be the case
if the cluster has reached its compute unit limit.