Request user to confirm password for aged sessions
An aged session is any session for which the SSO token was created more
than 24 hours ago.

When the user needs to confirm their password a modal dialog is shown.
What is displayed below the dialog depends on the state of the
application.

If the directory page has just been loaded, we display a spinner.  The
assumption being that the user will enter their password and the spinner
will continue to be displayed whilst loading any required details from
Flight Center.

If the confirm password modal has been dismissed, an message stating
that their password needs to be entered is shown.  Care is taken to 1)
not nag the user with another dialog and 2) avoid replacing the message
with a spinner when the user decides to show the dialog.

If the terminal was mounted prior to the dialog being shown we make sure
that it isn't unmounted, as we don't want to lose the session history.
If the dialog is dismissed, the terminal and the session history modal
if one is displayed are removed.